Move TimelineWriter into its own directory
